# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Educational Community Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses with optional filters
      parameters:
        - in: query
          name: categoryId
          schema:
            type: number
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
        - in: query
          name: difficulty
          schema:
            type: string
          required: false
        - in: query
          name: limit
          schema:
            type: number
          required: false
        - in: query
          name: offset
          schema:
            type: number
          required: false
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a draft course (editor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course by ID
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Course detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Replace course copy and withdraw it pending owner approval (editor)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "200":
          description: Draft course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}/approve:
    post:
      operationId: approveCourse
      tags: [courses]
      summary: Approve current course copy for publication (owner)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorialApproval"
      responses:
        "200":
          description: Published course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}/lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List lessons for a course
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lessons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Add a draft lesson to a course (editor)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a single lesson
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Replace lesson copy and withdraw it pending owner approval (editor)
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "200":
          description: Draft lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}/approve:
    post:
      operationId: approveLesson
      tags: [lessons]
      summary: Approve current lesson copy for publication (owner)
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditorialApproval"
      responses:
        "200":
          description: Published lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /editorial/courses/{courseId}:
    get:
      operationId: reviewCourse
      tags: [courses]
      summary: Review exact course draft and revision (editor or owner)
      parameters:
        - in: path
          name: courseId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Draft under review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorialCourse"
  /editorial/lessons/{lessonId}:
    get:
      operationId: reviewLesson
      tags: [lessons]
      summary: Review exact lesson draft and revision (editor or owner)
      parameters:
        - in: path
          name: lessonId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Draft under review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EditorialLesson"
...
components:
  schemas:
...
    Course:
      type: object
      required: [id, title, description, categoryId, categoryName, difficulty, instructorName, lessonCount, enrollmentCount, accessTier, createdAt]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        categoryId:
          type: number
        categoryName:
          type: string
        difficulty:
          type: string
        instructorName:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        lessonCount:
          type: number
        enrollmentCount:
          type: number
        accessTier:
          type: string
          enum: [Free, Elevated, Premium]
        transformationStory:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
    CourseDetail:
      type: object
      required: [id, title, description, categoryId, categoryName, difficulty, instructorName, lessonCount, enrollmentCount, accessTier, createdAt, lessons]
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        categoryId:
          type: number
        categoryName:
          type: string
        difficulty:
          type: string
        instructorName:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        lessonCount:
          type: number
        enrollmentCount:
          type: number
        accessTier:
          type: string
          enum: [Free, Elevated, Premium]
        transformationStory:
          type: ["string", "null"]
        isFeatured:
          type: boolean
        createdAt:
          type: string
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
    CourseInput:
      type: object
      required: [title, description, categoryId, difficulty, instructorName]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        categoryId:
          type: number
        difficulty:
          type: string
        instructorName:
          type: string
        thumbnailUrl:
          type: string
        isFeatured:
          type: boolean
        accessTier:
          type: string
          enum: [Free, Elevated, Premium]
        transformationStory:
          type: string
    EditorialApproval:
      type: object
      required: [revision]
      properties:
        revision:
          type: string
          minLength: 64
          maxLength: 64
    EditorialCourse:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          required: [revision, publishedAt]
          properties:
            revision: { type: string }
            publishedAt: { type: ["string", "null"] }
    EditorialLesson:
      allOf:
        - $ref: "#/components/schemas/Lesson"
        - type: object
          required: [revision, publishedAt]
          properties:
            revision: { type: string }
            publishedAt: { type: ["string", "null"] }
    Lesson:
      type: object
      required: [id, courseId, title, sortOrder, durationMinutes, createdAt]
      properties:
        id:
          type: number
        courseId:
          type: number
        title:
          type: string
        content:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        sortOrder:
          type: number
        durationMinutes:
          type: number
        createdAt:
          type: string
    LessonInput:
      type: object
      required: [title, sortOrder, durationMinutes]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        videoUrl:
          type: string
        sortOrder:
          type: number
        durationMinutes:
          type: number
...