# lib/api-zod/src/generated/types/editorialApproval.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Educational Community Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface EditorialApproval {
  /**
     * @minLength 64
     * @maxLength 64
     */
  revision: string;
}