# lib/api-zod/src/generated/types/editorialCourse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Educational Community Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from './course';

export type EditorialCourse = Course & ({
  revision: string;
  /** @nullable */
  publishedAt: string | null;
});