# lib/api-zod/src/generated/types/editorialLesson.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Educational Community Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Lesson } from './lesson';

export type EditorialLesson = Lesson & ({
  revision: string;
  /** @nullable */
  publishedAt: string | null;
});